requirements.txt
# Core Server
Flask==2.2.3
python-dotenv==1.0.0
Werkzeug<3
gunicorn==20.1.0  # For production deployment

# Microsoft Graph API / SharePoint
msal==1.24.1
requests==2.31.0

# Azure Key Vault
azure-keyvault-secrets==4.7.0
azure-identity==1.14.0

# Caching
redis==5.0.1

# Tools
python-docx==0.8.11
sqlparse==0.4.4
pandas==2.0.3
textstat==0.7.3

# Testing
pytest==7.4.2
pytest-mock==3.11.1